Step 05: separate classes
